[geoip] replace requirements.txt with Pipfile
Pipfile (used by pipenv) is the new and recommended way to handle python
dependencies. This change was brought on in part by a security warning from
github about a security issue in the version of the requests library that we
were using.

Pipfile makes it easier to upgrade to newer versions of the libraries used.

This change is reflected in the infrastructure repo as well, where pipenv is
used to install the geoip dependencies.