[dygraph-light] Fix preview legend order by associating legend labels with object ID (#343)
Provided in the search parameters are both `objId`, a comma-separated
list of object IDs, and `legendLabels` (and `legendLabelsY2`), a
corresponding list of legend labels for the primary Y-axis.

Internally, the data objects are re-ordered so that the data object with
the first recorded data (along the x-axis) is displayed first. However,
the legend labels were associated with the object ID, so they were
becoming decoupled during this reordering.

Thus, we now ensure that the legend label corresponding to the objectID
in question is displayed.

Also add appropriate Y2 labels when adding Y2 axis instead of leaving
them as file name, ensuring consistency with label names.

---
- To see the specific tasks where the Asana app for GitHub is being
used, see below:
  - https://app.asana.com/0/0/1209555804864695