[portal][imagezipview] Enable controlling imagezipview from portal (#344)
Let the user control the imagezipview preview from the keyboard without
needing to focus on imagezipview. Fixes bugs, handles missing images
gracefully, and improves consistency with the style of the imagezipview
iframe and the portal.

Details:

- Enables the portal to send keystrokes to the imagezipview iframe, so
controlling the images can be done using arrow keys without needing the
iframe to have focus. Adds WASD controls as well (A and D for
left/right).
- Changes the navigation controls for the images to be right-aligned
completely, making them more in line with the existing options for the
preview.
- Refactors `PreviewSelfContained.tsx`, changing it from a class-based
to function-based component. We should be able to perform these
refactors piecemeal in order to bring the React code up to a more modern
standard -- this is kind of the first example of doing this.
- Fixes a bug with the resize handler, where it would trigger too late
for map graph previews. (It is a modification of the existing strategy;
it would be better to use the `MutationObserver` API to watch the
iframe's `scrollHeight` and trigger an event when it changes, instead of
using a resize handler at all. This was beyond the scope of this PR,
though.)
- Adds error handling for missing images, ensuring an alt tag is applied
to inform the user that the image is missing, instead of just showing an
empty broken image icon.

_Note: While I tried to limit any stylistic code decisions, there are
definitely still some subjective changes to the code in here. We are
still working on establishing some kind of guidelines which will clarify
how to handle those, and they can be easily changed at that point._


---
- To see the specific tasks where the Asana app for GitHub is being
used, see below:
  - https://app.asana.com/0/0/1209593710989293