[preview] Remove scatter/line option for some datasets (#347)
Adds a config option to exclude the `ChartType` options (scatter or line
plot) from certain datasets, and hides it in those cases.

One dataset (`otcFosDataset`) uses the `TIMESERIES` preview type, but
has data incompatible with the current implementation of the tool.
Specifically, if both ocean and atmospheric measurements are present,
the rows are interspersed, featuring NaN values every other row. Due to
this (and the way Dygraphs handles NaN values as explicitly defined
gaps), the "line graph" views of these plots do not work consistently,
typically remaining as scatter plots.

This removes the non-working option from this dataset while we evaluate
what to do to solve the problem.