[preview][imagezipview] Provide persistent URL for specific image previews (#348)
Updates the iframe URL provided via the full-screen button and embed
links when the image is changed within the imagezipview previews
(multi-image zip).

Specifically, changes behavior to set the iframe src on the first
render, but not update it when the persistent src changes. The iframe
keeps its own state, and stays in sync with the portal by reporting the
img index back whenever it changes. The portal then uses that persistent
ID in providing the URL and embed iframe src link.

Without this change, the iframe would refresh every time the src is
changed as the component would re-render when the preview prop changed.

This change also fixes a bug with NetCDF previews, which are
experiencing that exact issue, where the iframe is refreshing every time
the src is changed.