[preview] Rename data cart caching, v1
In this version I changed SearchResultRegular to cache objects when the
checked objects array changes. This is honestly not a great way to do
it; instead, updateCheckedObjects should be updating the DataObject
cache to be accurate, which means I should do it globally, to maintain
consistency with current code. My approach is fine for handling state
more at the component level, but it's quite strange to mix this.
Unfortunately changing checkedObjectsInSearch in the global state to
contain DataObjects instead of strings is not easy; we would also have
to change checkedObjectsInCart and change a lot of functions that rely
on it being a string[], which is not ideal for just wanting to
intelligently rename the cart.