[preview] Improve URL hash synchronization (#350)
When a preview iframe changes, we ensure that the URL hash is updated so
that the URL for the preview page will generate the exact iframe shown
to the user. This way, users can share URLs with each other (or bookmark
a page and return to it) and ensure they will see the same preview.*

Adds explicit parameter checking when interacting with search or hash
parameters and fixes type declarations to match usage. Reduces history
pollution by using `replaceState` instead of `pushState` unless
explicitly needed/wanted.

*There are a handful of cases where there isn't true. Two cases are: 1)
parameters included in iframe URL but not respected on load (e.g. NetCDF
previews don't respect the `color` parameter); 2) interactive options
that are not included in iframe URL right now, but theoretically could
be (e.g. changes to the "range" options in NetCDF previews, or zooming
in or out on a time series Dygraph); in this case, we would need to
decide which ones should be included.