[preview] Fix TimeSeries y2 select default option (#353)
When selecting the parameter for y2 in time series previews and then
de-selecting it, the select menu and full-screen URL were not updating
properly. This is because we were storing the y2 setting in the
specSettings and then using it if the y2 parameter was unset. The
intention of this is to restore previously used selections when
returning to a specific time series spec preview, but because there is
currently no way to remove the y2 option when it is reset, this was not
functioning properly and was overriding the selection.

This change removes the y2 parameter from storage and defaults to
undefined if there is no value selected, which fixes the bug.