[portal] Remove previewSettings from state when leaving preview page (#354)
Ensure previewSettings are reset when leaving a preview page.

When navigating within the data portal, if you go to a preview and make
changes, then go back to search, then preview a different kind of data
object, those settings persist. This can lead to unexpected consequences
(e.g. a time series trying to show a variable that does not exist for
that data format).

This change removes previewSettings when navigating away from a
"preview" route page, ensuring it is always fresh when generating a new
preview from the search results page. Previously used settings persist
in the browser history, so a user could use the back arrows to go back
to an older preview (with settings) if they were trying to restore those
and accidentally navigated away.