[portal] Streamline KnownDataObject/ObjectsTable type (#364)
The type `ObjectsTable` is a constructed type that is currently not
obeyed by our runtime code; thus, it will be changed to reflect reality
and ensure type safety. Here, we redefine it to `KnownDataObject`, a
previously used type which is expanded to accomodate optional properties
and explicitly defined to prevent conflicts and `never` types.

Additionally, `ObjectsTable` is not a great name; a single
`ObjectsTable` is not a table of multiple objects, but instead the
information about a single object. Note that an array of
`KnownDataObjects` is still often known as an `objectsTable`; this makes
sense, as it is a table of known data objects, but this could be changed
if it is unclear.

More details in PR.