[portal] Empty cart after download (#373)
Add automatic cart emptying (and potential cart restoration, if
necessary) after downloading data.

Note that the DataCart container rendering was restructured; this is
necessary in order to ensure the `form` always exists in the DOM and has
the correct information. With conditional rendering, the form can become
disconnected prior to the completion of the `submit` event, which leaves
the "side effects" of the form submission without actually submitting
the form (and thus triggering the download).

`actions/cart.ts` was also changed to use the `updateCart` function from
`actions/common.ts`. These two functions are nearly identical, but the
`common.ts` function also updates the number in the cart at the top
right of the screen, using it instead ensures everything stays in sync.