[portal] Rename data cart intelligently (#374)
Changes the name of the downloaded file link from "My data" to be based
on timestamp, station (if all the same), and spec (if all the same).

Introduces a new directory for `hooks`, custom functions similar to
`useState`, `useEffect`, or `useMemo`. This enables reusable hooks to
help manage more complex local state and provide the results across
multiple components, the first being `useDownloadInfo` which provides
information about the objects ready for download (in this case,
filename, but could be expanded as necessary).

It might have been more natural to put this into the `Cart` model
itself, but since we do not use the `Cart` when checking objects on the
main search page for download, the file name generation needed to be
possible in both contexts. The filename generation is more difficult for
the search, which is why a hook is used: results are refreshed when the
page changes, but the current results (for checked data objects) need to
be cached so that the `KnownDataObject` fields are available for the
filename generation.

Thus, the broadest use case forms the basis for `useDownloadInfo`, with
the narrower use case of the `DataCart` container/interface fitting into
it by providing its object info via a `localObjectsTable` instead of a
global `objectsTable`. The caching is, of course, unnecessary for the
`DataCart`; but the hook and its use of `useMemo` ensures this overhead
does not have an impact.

(A few changes were also made to remove unnecessary dependencies that
were noticed and missed from PR #365 )

---------

Co-authored-by: Jonathan Thiry <github@jonathanthiry.com>