[portal] Ensure undefined parameters are not passed to NetCDF previews (#386)
In PR #359 we explicitly set preview settings to `undefined`, but we did
not filter them out when using query parameters in NetCDF previews.
MapGraph worked fine, because the hash setting with `JSON.stringify`
ignores `undefined` values, but when the settings become query
parameters, the `undefined` values are implicitly converted into
strings, so they need to be explicitly filtered out.