[portal] Always create new CartItem for Previews (#400)
Fixes bug where `PreviewTitle` does not render when previewed dobj is in
the cart.

Essentially, when the cart is restored, it does not fully construct the
`CartItem`s. Instead, the data is fetched when needed; but it does not
restore the preview type. This, then, prevents the `PreviewTitle` from
being rendered, since it checks for the item having a preview type (i.e.
being "previewable") to determine whether or not to show the
`PreviewTitle`. It should be only a slight amount of overhead to create
a new `CartItem` when initializing a `Preview` which contains the proper
preview type, since the data is all already fetched inside
`initPreview`.

Note that the cause of the bug is when `PreviewTitle.tsx` encounters an
item with an `undefined` type:
<https://github.com/ICOS-Carbon-Portal/data/blob/67e7de96d6fb870754446e4d97d7e7bb7055f802/src/main/js/portal/main/components/preview/PreviewTitle.tsx#L19>