Use IP2Location database instead of IPstack API for geolocation of downloaders' IP addresses (#404)
Some of the information that was obtained through IPstack is not
available with IP2Location and is now ignored (continent name, continent
code, region code, ZIP code and whether the location is within an EU
country).
The SQLite cache has been removed because it is not relevant anymore
with the new solution.

---------

Co-authored-by: Jonathan Thiry <github@jonathanthiry.com>